simplify vulnerability parser; organize packages
