+++ go.mod
@@ -1,0 +1,1 @@
+module github.com/billiford/slingshot
